fix: 스프라이트 파티클 생성 시 메시 파티클도 같이 생성되는 문제 수정

+++ EngineSIU/EngineSIU/Engine/Source/Runtime/Engine/ParticleHelper.cpp
@@ -43,7 +43,7 @@ void FParticleDataContainer::Free()
 
 void FDynamicEmitterReplayDataBase::Serialize(FArchive& Ar)
 {
-    int32 EmitterTypeAsInt = eEmitterType;
+    int32 EmitterTypeAsInt = static_cast<int32>(eEmitterType);
     Ar << EmitterTypeAsInt;
     eEmitterType = static_cast< EDynamicEmitterType >( EmitterTypeAsInt );
